requirements.txt
# Python backend dependencies
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
pydantic==2.11.7
pydantic-ai==0.8.1
sqlalchemy==2.0.43
alembic==1.16.4
asyncpg==0.30.0
greenlet==3.2.4
loguru==0.7.3
python-dotenv==1.1.1
PyYAML==6.0.2
openai==1.107.1
pinecone==7.3.0
httpx==0.28.1
sse-starlette==3.0.2

# LangChain/LangGraph for advanced agent capabilities
langchain==0.3.27
langchain-openai==0.3.33
langchain-community==0.3.29
langgraph==0.6.7
